+++ display/subscriptions.go
@@ -97,9 +97,11 @@ func Subscriptions(t *tab, u string) string {
 	} else {
 		// Render page
 
-		rawPage += "You can use Ctrl-X to subscribe to a page, or to an Atom/RSS/JSON feed. See the online wiki for more.\n" +
-			"If you just opened Amfora then updates may appear incrementally. Reload the page to see them.\n\n" +
-			"=> about:manage-subscriptions Manage subscriptions\n\n"
+		if viper.GetBool("subscriptions.header") {
+			rawPage += "You can use Ctrl-X to subscribe to a page, or to an Atom/RSS/JSON feed. See the online wiki for more.\n" +
+				"If you just opened Amfora then updates may appear incrementally. Reload the page to see them.\n\n"
+		}
+		rawPage += "=> about:manage-subscriptions Manage subscriptions\n\n"
 
 		// curDay represents what day of posts the loop is on.
 		// It only goes backwards in time.